requirements.txt
requests~=2.31.0
python-dotenv~=1.0.0
python-okx~=0.2.0
